requirements.txt
# SparkJAR Shared Package Requirements
# Minimal dependencies - most come from parent packages

# Core dependencies that tools/utils need
pydantic>=2.0.0
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Authentication
pyjwt>=2.9.0

# LLM clients
openai>=1.13.3

# Document processing
pypdf>=3.17.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# Google integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1

# CrewAI tools support
crewai>=0.148.0
